requirements.txt
Django==1.5.2
South==0.8.2
dj-database-url==0.2.2
dj-static==0.0.5
django-discover-runner==1.0
django-registration==1.0
django-toolbelt==0.0.1
djangorestframework==2.3.8
facebook-sdk==0.4.0
gunicorn==18.0
psycopg2==2.5.1
static==0.4
validate-email==1.1
wsgiref==0.1.2
